+++ pkg/core/certifier/challengers/dns01/azure/azure.go
@@ -1,4 +1,4 @@
-package azuredns
+package azure
 
 import (
 	"fmt"
