+++ pkg/core/certifier/challengers/dns01/googlecloud/googlecloud.go
@@ -1,4 +1,4 @@
-package googleclouddns
+package googlecloud
 
 import (
 	"context"
