+++ pkg/core/certmgr/providers/ksyun-kcm/ksyun_kcm.go
@@ -137,8 +137,10 @@ func (c *Certmgr) tryGetResultIfCertExists(ctx context.Context, certPEM string) 
 			break
 		}
 
+		fingerprintSha1 := sha1.Sum(certX509.Raw)
+		fingerprintSha1Hex := hex.EncodeToString(fingerprintSha1[:])
 		for _, certItem := range listUserCertificatesResp.Ret.Certs {
-			// 对比证书多域名
+			// 对比证书备用名称
 			if !strings.EqualFold(strings.Join(certX509.DNSNames, ","), strings.Join(certItem.Domains, ",")) {
 				continue
 			}
@@ -149,9 +151,7 @@ func (c *Certmgr) tryGetResultIfCertExists(ctx context.Context, certPEM string) 
 			}
 
 			// 对比证书指纹
-			fingerprint := sha1.Sum(certX509.Raw)
-			fingerprintHex := hex.EncodeToString(fingerprint[:])
-			if !strings.EqualFold(fingerprintHex, certItem.FingerPrint) {
+			if !strings.EqualFold(fingerprintSha1Hex, certItem.FingerPrint) {
 				continue
 			}
