+++ pkg/core/certmgr/providers/wangsu-certificate/wangsu_certificate.go
@@ -109,7 +109,7 @@ func (c *Certmgr) Upload(ctx context.Context, certPEM, privkeyPEM string) (*Uplo
 		Name:        lo.ToPtr(certName),
 		Certificate: lo.ToPtr(certPEM),
 		PrivateKey:  lo.ToPtr(privkeyPEM),
-		Comment:     lo.ToPtr("upload from certimate"),
+		Comment:     lo.ToPtr("upload from Certimate"),
 	}
 	createCertificateResp, err := c.sdkClient.CreateCertificateWithContext(ctx, createCertificateReq)
 	c.logger.Debug("sdk request 'certificatemanagement.CreateCertificate'", slog.Any("request", createCertificateReq), slog.Any("response", createCertificateResp))
@@ -141,7 +141,7 @@ func (c *Certmgr) Replace(ctx context.Context, certIdOrName string, certPEM, pri
 		Name:        lo.ToPtr(certName),
 		Certificate: lo.ToPtr(certPEM),
 		PrivateKey:  lo.ToPtr(privkeyPEM),
-		Comment:     lo.ToPtr("upload from certimate"),
+		Comment:     lo.ToPtr("upload from Certimate"),
 	}
 	updateCertificateResp, err := c.sdkClient.UpdateCertificateWithContext(ctx, certId, updateCertificateReq)
 	c.logger.Debug("sdk request 'certificatemanagement.UpdateCertificate'", slog.Any("request", updateCertificateReq), slog.Any("response", updateCertificateResp))
